Moved responsibility for managing Eclipse dependencies back in Eclipse.
This means that the plugins must be built by a headless eclipse.  The problem is that dependencies required by the editor sub-plugin appear to not be present in the mvn repository (at least not the most recent versions) making dpendency management via ivy difficult.  There appears to be Ant tasks for the P2 OSGI provisioning system that may allow jar files to be pulled as bundles.

git-svn-id: file:///extra/svn-repositories/resme/branches/ant-build-script@82 13975f06-6b5f-4edb-8417-069f660ee22b